requirements.txt
google-cloud==0.34.0